// src/generated/models/Cre2b_servicecenterproximitiesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cre2b_servicecenterproximitiesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cre2b_servicecenterproximitiesstatecode = keyof typeof Cre2b_servicecenterproximitiesstatecode;
export const Cre2b_servicecenterproximitiesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cre2b_servicecenterproximitiesstatuscode = keyof typeof Cre2b_servicecenterproximitiesstatuscode;

export interface Cre2b_servicecenterproximitiesBase {
  "cre2b_Closest1TSRAdvisor@odata.bind"?: string;
  "cre2b_Closest2TSRAdvisor@odata.bind"?: string;
  "cre2b_Closest3TSRAdvisor@odata.bind"?: string;
  cre2b_closesttsr1?: string;
  "cre2b_ClosestTSR1TSRonly@odata.bind"?: string;
  cre2b_closesttsr2?: string;
  "cre2b_ClosestTSR2TSRonly@odata.bind"?: string;
  cre2b_closesttsr3?: string;
  "cre2b_ClosestTSR3TSRonly@odata.bind"?: string;
  cre2b_closesttsroradvisor1?: string;
  cre2b_closesttsroradvisor2?: string;
  cre2b_closesttsroradvisor3?: string;
  cre2b_distancetoclosesttsroradvisor1?: number;
  cre2b_distancetoclosesttsroradvisor2?: number;
  cre2b_distancetoclosesttsroradvisor3?: number;
  cre2b_distancetotsr1?: number;
  cre2b_distancetotsr2?: number;
  cre2b_distancetotsr3?: number;
  cre2b_proximityname?: string;
  cre2b_servicecenter?: string;
  "cre2b_ServiceCenterLookup@odata.bind"?: string;
  cre2b_servicecenterproximityid: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Cre2b_servicecenterproximitiesstatecode;
  statuscode?: Cre2b_servicecenterproximitiesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cre2b_servicecenterproximities extends Cre2b_servicecenterproximitiesBase {
  cre2b_closest1tsradvisorname?: string;
  cre2b_closest2tsradvisorname?: string;
  cre2b_closest3tsradvisorname?: string;
  cre2b_closesttsr1tsronlyname?: string;
  cre2b_closesttsr2tsronlyname?: string;
  cre2b_closesttsr3tsronlyname?: string;
  cre2b_servicecenterlookupname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  cre2b_closest1tsradvisor?: object;
  _cre2b_closest1tsradvisor_value?: string;
  cre2b_closest2tsradvisor?: object;
  _cre2b_closest2tsradvisor_value?: string;
  cre2b_closest3tsradvisor?: object;
  _cre2b_closest3tsradvisor_value?: string;
  cre2b_closesttsr1tsronly?: object;
  _cre2b_closesttsr1tsronly_value?: string;
  cre2b_closesttsr2tsronly?: object;
  _cre2b_closesttsr2tsronly_value?: string;
  cre2b_closesttsr3tsronly?: object;
  _cre2b_closesttsr3tsronly_value?: string;
  cre2b_servicecenterlookup?: object;
  _cre2b_servicecenterlookup_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}